Implemented the lexer.
There is now a basic lexer in place that recognizes integers. I've also
added the beginnings of a REPL. For now it just reads in lines of input
and displays the resulting tokens.

A new EOL (end of line) token type was added, which is just an ad hoc
way of letting the driver know that an entire line has been read. When
we have a parser in place, newlines will probably be ignored just like
all other kinds of whitespace.

The next step is getting a parser (Scheme's 'read' procedure) written,
along with an expression evaluator and printer. Then we can make a
real REPL and start adding support for more expression types.