Started the object model.
The parser will take tokens and convert them into objects, which are
nodes that make up the syntax tree, which is what the evaluator will
work on. Right now, the only kind of object is the number.

Our object model is more or less identical to Peter Michaux's model
in his Bootstrap Scheme, although our parser is structured differently.