read_line() now handles arbitrarily long lines.
Previously, read_line() used fgets() to read an entire lines at a time,
and stored them in a large fixed-sized buffer. The new version
dynamically allocates a buffer, and resizes it as needed. As a bonus, it
also stops reading when it encounters a null character, unlike fgets.
There is a getline() function in glibc which does nearly what I want,
but I'd prefer to keep the code portable to other systems.

The lex_* functions still work in terms of array indices, and will need
to be cleaned up a bit.