Prototyped support for strings.
The lex_string() function in lexer.c is just stubbed out. Finishing it
will require parsing the string, looking for escape sequences and
replacing them with the corresponding character. Also, allocating a
buffer to store the string. The sb_write() function will also have to be
updated to translate the characters back into escape sequences.

For now, we'll just support three sequences: \n for newline, \" for a
'"' character, and \\ for a '\' character. Embedded literal newlines
are not allowed.