Added the quote special form.
This is the first step away from being just a pretty-printer. Symbols
and pairs are no longer self-evaluating, but must be quoted. Next is
support for environments, and the define and set! special forms. Most of
the action will be in the evaluator now, instead of the lexer and
parser.