Moved (), #t, and #f from parser.c to object.c
It makes more sense for them to be there, and will come in handy when
environments are added to eval.c, and we'll need access to the empty
list. It also makes bs_read() a little more regular.

There are accessor functions (get_empty_list() and get_boolean(value))
for retrieving the objects.