Implemented environments and variables.
Simple definitions and setting variables work. (define and set! in
Scheme). No procedures yet, because we don't have a way of representing
them.

The code in eval.c leans pretty heavily on Peter Michaux's
bootstrap-scheme code, which is in turn a pretty faithful translation to
C of the environment procedures and data structures from chapter 4 of
SICP.

References:
http://peter.michaux.ca/articles/scheme-from-scratch-bootstrap-v0_9-environments
http://mitpress.mit.edu/sicp/full-text/book/book-Z-H-26.html#%_sec_4.1.3