Added support for 'if' expressions.
The result of evaluating the expression (if #f 'a) is unspecified by the
standard, but it does not disallow 'if' expressions with no alternate
expression. My first inclination would be to make such an expression an
error, but we'll follow boostrap-scheme (and SICP) here, and return #f.

If it's good enough for Abelson and Sussman, it's good enough for me.