Factored out the environment code to its own file.
All the environment handling is in environment.c now. I've also switched
to the model described exercise 4.11 of SICP. All tests pass, so I'm
committing the changes, but I'm going to continue cleaning up the
code base. There've been a lot of commits in the last 11 days, and some
things aren't written as clearly as they could be.