Added symbol->string and string->symbol
string->symbol can create symbols that don't follow that lexical rules
for symbols, so that's something to look out for.

I also found a bug in the lex_string that makes it choke on strings that
start with an escape sequence, i.e. "\n", but not " \n". I'll fix it as
soon as I figure out why it's happening.