Added 'load' procedure.
To go along with that, there is an init.scm file that defines a few
useful procedures.

Input file "handling" was pulled out of the lexer and put into it's own
file. A better design would be for set_input_file and get_input_file to
take and return strings, and do the file opening and closing themselves.