Make user errors a little less fatal.
Most errors no longer cause the interpreter to exit. Ideally, the only
fatal errors should be due to things like running out of memory, or
specifying a non-existant file on the command line. In practice, some
lexical errors will also cause the interpreter to give up.

Functions the would have previously exited on an error now print a
warning message and, if necessary, return an object of type 'INVALID'.
That object indicates that the previous action failed, and its result
should be ignored.

I would like to make error messages a little more useful to the user by
giving some context in terms of the Scheme input, rather than the C
internals, but I think we can call this "good enough" for now.