Added 'string-append', and fixed some bugs in I/O.
I'm reading "Lisp In Small Pieces", and wanted to implement some of the
interpreters from that book in bs. That means I need 'gensym', and
'string-append' will be useful for writing it.

I ran into a couple bug while testing out 'string-append':

The standard ports were being initialized last. If any error messages
are printed before that, it would cause a segfault. In this case, I had
turned on INFO messages, which the symbol table prints when a
hash-collision occurs. Initializing the ports first fixed the bug.

I also realized that if the current error port is closed, calling
error() to print a message about it won't work. So instead, we just
use fprintf(stderr, ...) and exit(); stderr should always be open.